requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
Flask-Bootstrap==3.3.7.0
Flask-Paginate==0.5.1
python-dotenv==0.9.1